+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Added new `--min-monthly-fee` and `--max-monthly-fee` options to filter flats by
+  monthly fee. The default is no minimum or maximum monthly fee.
 - Added new `--cache/--no-cache` flag, which allows you to disable the cache. The
   default behaviour is still to use the cache, but you can disable it by using the
   `--no-cache` flag. This is useful if you want to see all the results, and not just the